+++ src/utils/OverlayServiceModule.ts
@@ -5,6 +5,9 @@ interface OverlayServiceModuleType {
   stopOverlayService(): Promise<boolean>;
   setButtonOpacity(opacity: number): Promise<boolean>;
   getButtonOpacity(): Promise<number>;
+  setStatusBarEnabled(enabled: boolean): Promise<boolean>;
+  getStatusBarEnabled(): Promise<boolean>;
+  setTestMode(enabled: boolean): Promise<boolean>;
 }
 
 const OverlayServiceModule: OverlayServiceModuleType = NativeModules.OverlayServiceModule;